update sprites and saved history

+++ src/server/components/utils/files.js
@@ -45,16 +45,38 @@ function update(file, data, cb) {
 }
 
 function move(oldPath, newPath, cb) {
-  fs.rename(
-    path.join(config.FILES_PATH, oldPath),
-    path.join(config.FILES_PATH, newPath),
-    err => cb(response.commonResult(err, newPath))
-  );
+  newPath = path.join(config.FILES_PATH, newPath);
+
+  function move(result) {
+    if (result && result.code !== 0 && result.description && result.description.code !== 'EEXIST' ) {
+      return cb(result);
+    }
+    fs.rename(
+      path.join(config.FILES_PATH, oldPath),
+      newPath,
+      onMove
+    );
+  }
+  function onMove(err) {
+    if (err && err.code == 'ENOENT') {
+      return mkdir(path.dirname(newPath), move);
+    }
+    cb(response.commonResult(err, newPath));
+  }
+  move();
 }
 
+function remove(file, cb) {
+  var file = path.join(config.FILES_PATH, file);
+  fs.unlink(
+    file,
+    err => cb(response.commonResult(err, file))
+  );
+}
 
 exports.move = move;
 exports.write = write;
 exports.update = write;
 exports.join = path.join;
 exports.mkdir = mkdir;
+exports.remove = remove;
